Add mmap_windows.go
